+++ pypose/__init__.py
@@ -5,12 +5,12 @@ from .lietensor import randn_like, randn_SE3, randn_SO3, randn_so3, randn_se3
 from .lietensor import randn_Sim3, randn_sim3, randn_RxSO3, randn_rxso3
 from .lietensor import identity_like, identity_SO3, identity_so3, identity_SE3, identity_se3
 from .lietensor import identity_Sim3, identity_sim3, identity_RxSO3, identity_rxso3
-from .lietensor import add, add_, mul, cumops, cummul, cumprod, cumops_, cummul_, cumprod_, pm
-from .lietensor import Exp, Log, Inv, Mul, Retr, Act, Adj, AdjT, Jinvp, Jr
+from .lietensor import add, add_, mul, Exp, Log, Inv, Mul, Retr, Act, Adj, AdjT, Jinvp, Jr
 from .lietensor import SO3_type, so3_type, SE3_type, se3_type
 from .lietensor import Sim3_type, sim3_type, RxSO3_type, rxso3_type
 from .lietensor import tensor, translation, rotation, scale, matrix, euler
 from .lietensor import mat2SO3, mat2SE3, mat2Sim3, mat2RxSO3, from_matrix, matrix, euler2SO3, vec2skew
 from .lietensor import gradcheck, gradgradcheck
+from .lietensor.function import *
 from . import module, optim
 from .basics import *
