+++ pypose/lietensor/__init__.py
@@ -8,8 +8,7 @@ from .utils import identity_like, identity_SO3, identity_so3, identity_SE3, iden
 from .utils import identity_Sim3, identity_sim3, identity_RxSO3, identity_rxso3
 from .utils import SO3, so3, SE3, se3, Sim3, sim3, RxSO3, rxso3
 from .utils import Exp, Log, Inv, Mul, Retr, Act, Adj, AdjT, Jinvp, Jr
-from .basics import vec2skew, cumops, cummul, cumprod
+from .basics import vec2skew, add, add_, mul
 from .convert import tensor, translation, rotation, scale, matrix, euler
-from .basics import add, add_, mul, cumops_, cummul_, cumprod_, pm
 from .convert import mat2SO3, mat2SE3, mat2Sim3, mat2RxSO3, from_matrix, matrix, euler2SO3
 from .gradcheck import gradcheck, gradgradcheck
